backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["RWallan <3am.richardwallan@gmail.com>"]
readme = "README.md"
packages = [{ include = "backend" }]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

[tool.poetry.urls]
"Documentation" = "https://github.com/RWallan/track-my-attendance/blob/main/README.md"
"Code" = "https://github.com/RWallan/track-my-attendance"
"Bug Tracker" = "https://github.com/RWallan/track-my-attendance/issues"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.1"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
pydantic-settings = "^2.5.2"
pydantic = {extras = ["email"], version = "^2.9.1"}
pwdlib = {extras = ["argon2"], version = "^0.2.1"}


[tool.poetry.group.dev.dependencies]
taskipy = "^1.13.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
ruff = "^0.6.4"
httpx = "^0.27.2"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = "ruff check .; ruff check . --diff"
format = "ruff check . --fix; ruff format ."
run = "fastapi dev backend/app.py"
pre_test = "task lint"
test = "pytest -s -x --cov=backend -vv"
post_test = "coverage html"
